src/requirements.txt
google-cloud-bigquery
jsonschema
pyyaml